Added per-server view of results.
The full interop table is very confusing: a lot of abbreviations wihout
explanation on same page, and is not very practical, when you want to
track specific implementation.

The patch adds one more table, that shows how different clients interact
with selected server.  Just click on the server name in the table head,
and the table below will be populated.  The left part of the table
contains non-abbreviated feature names.