Fix flaky compliance checks
Previously, the compliance check seeks for the strings "exited with
code 127" or "exit status 127" in stdout + stderr.  It turns out that
these strings might not be present in rare cases.  In order to
workaround this, this commit directly checks the exit code of the
relevant container with docker-compose --exit-code-from flag.  The
flag implies --abort-on-container-exit.

When running the actual test case, the interop runner checks whether
the test case is supported by an implementation.  The same method
cannot be applied there because we only get an exit code from a single
service.  However, the downside of not detecting unsupported test case
is not severe, it just results in failed test.  In contrast, the
failed compliance check skips all test cases for the particular client
and server combination.