Don't check for retry in handshake test
The test checks for retries, but according to RFC 900 s. 8.1, a server
may send a retry packet to preform address validation on the client.  If
the server does this the trace will show a retry packet, and doing so
appears valid

easiest fix seems to simply be to allow retry packets to be present in
the test